requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
livekit-server-sdk-python==1.0.5
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.7
Werkzeug==2.2.3
openai==1.3.0
pyngrok==6.0.0